moved filesystem inside pkg dir
